Bring Report templates project back into v13 changes and update references to bema version.
