+++ rocks.kfs.RoomManagement.Templates/LavaLandscapeReportTemplate.cs
@@ -29,8 +29,8 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.IO;
 using System.Linq;
-using com.centralaz.RoomManagement.Model;
-using com.centralaz.RoomManagement.ReportTemplates;
+using com.bemaservices.RoomManagement.Model;
+using com.bemaservices.RoomManagement.ReportTemplates;
 using iTextSharp.text;
 using iTextSharp.text.html.simpleparser;
 using iTextSharp.text.pdf;
