+++ rocks.kfs.RoomManagement.Templates/ResourceQuestionReportTemplate.cs
@@ -20,8 +20,7 @@
 // and is a derivative work.
 //
 // Modification (including but not limited to):
-// * Added Ministry and Contact Info to columns.
-// * Reformatted columns for client needs.
+// * Added new section to repeat through resource questions
 // </notice>
 //
 using System;
@@ -29,27 +28,23 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.IO;
 using System.Linq;
-using com.centralaz.RoomManagement.Attribute;
-using com.centralaz.RoomManagement.Model;
-using com.centralaz.RoomManagement.ReportTemplates;
+using com.bemaservices.RoomManagement.Attribute;
+using com.bemaservices.RoomManagement.Model;
+using com.bemaservices.RoomManagement.ReportTemplates;
 using iTextSharp.text;
 using iTextSharp.text.pdf;
 using Rock;
-using Rock.Data;
-using Rock.Model;
-using Document = iTextSharp.text.Document;
 
 namespace rocks.kfs.RoomManagement.ReportTemplates
 {
     /// <summary>
-    /// 
+    ///
     /// </summary>
-    [System.ComponentModel.Description( "The SPAC advanced report template" )]
+    [System.ComponentModel.Description( "Resource question report template" )]
     [Export( typeof( ReportTemplate ) )]
-    [ExportMetadata( "ComponentName", "SPAC" )]
-    public class SPACReportTemplate : ReportTemplate
+    [ExportMetadata( "ComponentName", "Resource Questions" )]
+    public class ResourceQuestionReportTemplate : ReportTemplate
     {
-
         /// <summary>
         /// Gets or sets the exceptions.
         /// </summary>
@@ -73,6 +68,7 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
             var titleFont = FontFactory.GetFont( font, 16, Font.BOLD );
             var listHeaderFont = FontFactory.GetFont( font, 12, Font.BOLD, Color.DARK_GRAY );
             var listSubHeaderFont = FontFactory.GetFont( font, 10, Font.BOLD, Color.DARK_GRAY );
+            var listItemFontBold = FontFactory.GetFont( font, 8, Font.BOLD );
             var listItemFontNormal = FontFactory.GetFont( font, 8, Font.NORMAL );
             var listItemFontUnapproved = FontFactory.GetFont( font, 8, Font.ITALIC, Color.MAGENTA );
             var noteFont = FontFactory.GetFont( font, 8, Font.NORMAL, Color.GRAY );
@@ -102,15 +98,14 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
             .Select( r => r.ToList() )
             .ToList();
 
-
             //Setup the document
-            var document = new Document( PageSize.A4.Rotate(), 25, 25, 25, 25 );
+            var document = new Document( PageSize.LETTER.Rotate(), 25, 25, 25, 50 );
 
             var outputStream = new MemoryStream();
             var writer = PdfWriter.GetInstance( document, outputStream );
 
             // Our custom Header and Footer is done using Event Handler
-            SPACTwoColumnHeaderFooter PageEventHandler = new SPACTwoColumnHeaderFooter();
+            ResourceQuestionHeaderFooter PageEventHandler = new ResourceQuestionHeaderFooter();
             writer.PageEvent = PageEventHandler;
 
             // Define the page header
@@ -140,7 +135,7 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
 
             Font zapfdingbats = new Font( Font.ZAPFDINGBATS );
 
-            // Populate the Lists            
+            // Populate the Lists
             foreach ( var reservationDay in reservationSummaries )
             {
                 var firstReservation = reservationDay.FirstOrDefault();
@@ -154,7 +149,7 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
                     //Build Subheaders
                     var listSubHeaderTable = new PdfPTable( 8 );
                     listSubHeaderTable.LockedWidth = true;
-                    listSubHeaderTable.TotalWidth = PageSize.A4.Rotate().Width - document.LeftMargin - document.RightMargin;
+                    listSubHeaderTable.TotalWidth = PageSize.LETTER.Rotate().Width - document.LeftMargin - document.RightMargin;
                     listSubHeaderTable.HorizontalAlignment = 0;
                     listSubHeaderTable.SpacingBefore = 10;
                     listSubHeaderTable.SpacingAfter = 0;
@@ -165,7 +160,7 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
                     listSubHeaderTable.AddCell( new Phrase( "Name", listSubHeaderFont ) );
                     listSubHeaderTable.AddCell( new Phrase( "Event Time", listSubHeaderFont ) );
                     listSubHeaderTable.AddCell( new Phrase( "Reservation Time", listSubHeaderFont ) );
-                    listSubHeaderTable.AddCell( new Phrase( "Locations: NC, Unlock", listSubHeaderFont ) );
+                    listSubHeaderTable.AddCell( new Phrase( "Locations", listSubHeaderFont ) );
                     listSubHeaderTable.AddCell( new Phrase( "Resources", listSubHeaderFont ) );
                     listSubHeaderTable.AddCell( new Phrase( "Has Layout", listSubHeaderFont ) );
                     listSubHeaderTable.AddCell( new Phrase( "Status", listSubHeaderFont ) );
@@ -179,11 +174,19 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
                         {
                             PageEventHandler.IsHeaderShown = false;
                         }
+                        var keepTogetherEvent = new PdfPTable( 1 );
+                        keepTogetherEvent.LockedWidth = true;
+                        keepTogetherEvent.TotalWidth = PageSize.LETTER.Rotate().Width - document.LeftMargin - document.RightMargin;
+                        keepTogetherEvent.KeepTogether = true;
+                        keepTogetherEvent.HorizontalAlignment = 0;
+                        keepTogetherEvent.SpacingBefore = 0;
+                        keepTogetherEvent.SpacingAfter = 1;
+                        keepTogetherEvent.DefaultCell.BorderWidth = 0;
 
                         //Build the list item table
                         var listItemTable = new PdfPTable( 8 );
                         listItemTable.LockedWidth = true;
-                        listItemTable.TotalWidth = PageSize.A4.Rotate().Width - document.LeftMargin - document.RightMargin;
+                        listItemTable.TotalWidth = PageSize.LETTER.Rotate().Width - document.LeftMargin - document.RightMargin;
                         listItemTable.HorizontalAlignment = 0;
                         listItemTable.SpacingBefore = 0;
                         listItemTable.SpacingAfter = 1;
@@ -194,7 +197,6 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
                         //    listItemTable.DefaultCell.BorderWidthBottom = 1;
                         //    listItemTable.DefaultCell.BorderColorBottom = Color.DARK_GRAY;
                         //}
-
 
                         //Add the list items
                         listItemTable.AddCell( new Phrase( reservationSummary.ReservationName, listItemFontNormal ) );
@@ -232,8 +234,7 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
                         resourceList.SetListSymbol( "\u2022" );
                         foreach ( var reservationResource in reservationSummary.Resources )
                         {
-                            reservationResource.LoadReservationResourceAttributes();
-                            var resourceListItem = new iTextSharp.text.ListItem( string.Format("{0} ({1})", reservationResource.Resource.Name, reservationResource.Resource.Quantity), listItemFontNormal );
+                            var resourceListItem = new iTextSharp.text.ListItem( string.Format( "{0} ({1})", reservationResource.Resource.Name, reservationResource.Quantity ), listItemFontNormal );
                             resourceList.Add( resourceListItem );
                         }
 
@@ -267,28 +268,132 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
                         //}
                         listItemTable.AddCell( contactCell );
 
-                        document.Add( listItemTable );
+                        //document.Add( listItemTable );
+
+                        PdfPCell eventRow = new PdfPCell();
+                        eventRow.Border = 0;
+                        eventRow.AddElement( listItemTable );
+                        keepTogetherEvent.AddCell( eventRow );
 
                         if ( !string.IsNullOrWhiteSpace( reservationSummary.Note ) )
                         {
                             //document.Add( Chunk.NEWLINE );
                             var listNoteTable = new PdfPTable( 8 );
                             listNoteTable.LockedWidth = true;
-                            listNoteTable.TotalWidth = PageSize.A4.Rotate().Width - document.LeftMargin - document.RightMargin;
+                            listNoteTable.TotalWidth = PageSize.LETTER.Rotate().Width - document.LeftMargin - document.RightMargin;
                             listNoteTable.HorizontalAlignment = 1;
                             listNoteTable.SpacingBefore = 0;
                             listNoteTable.SpacingAfter = 1;
                             listNoteTable.DefaultCell.BorderWidth = 0;
-                            listNoteTable.DefaultCell.BorderWidthBottom = 1;
-                            listNoteTable.DefaultCell.BorderColorBottom = Color.DARK_GRAY;
+                            //listNoteTable.DefaultCell.BorderWidthBottom = 1;
+                            //listNoteTable.DefaultCell.BorderColorBottom = Color.DARK_GRAY;
                             listNoteTable.AddCell( new Phrase( string.Empty, noteFont ) );
                             var noteCell = new PdfPCell( new Phrase( reservationSummary.Note, noteFont ) );
                             noteCell.Border = 0;
                             noteCell.Colspan = 7;
                             listNoteTable.AddCell( noteCell );
 
-                            document.Add( listNoteTable );
+                            //document.Add( listNoteTable );
+
+                            PdfPCell noteRow = new PdfPCell();
+                            noteRow.Border = 0;
+                            noteRow.AddElement( listNoteTable );
+                            keepTogetherEvent.AddCell( noteRow );
                         }
+
+                        if ( reservationSummary.Resources.Any() )
+                        {
+                            var listResourceTable = new PdfPTable( 8 );
+                            listResourceTable.LockedWidth = true;
+                            listResourceTable.TotalWidth = PageSize.LETTER.Rotate().Width - document.LeftMargin - document.RightMargin;
+                            listResourceTable.HorizontalAlignment = 1;
+                            listResourceTable.SpacingBefore = 0;
+                            listResourceTable.SpacingAfter = 1;
+                            listResourceTable.DefaultCell.BorderWidth = 0;
+                            listResourceTable.DefaultCell.BorderWidthBottom = 1;
+                            listResourceTable.DefaultCell.BorderColorBottom = Color.DARK_GRAY;
+                            listResourceTable.KeepTogether = true;
+
+                            var addResourceTable = false;
+                            foreach ( var reservationResource in reservationSummary.Resources )
+                            {
+                                reservationResource.LoadReservationResourceAttributes();
+
+                                if ( reservationResource.Attributes != null && reservationResource.Attributes.Any() )
+                                {
+                                    var headerCell = new PdfPCell( new Phrase( string.Format( "{0} ({1})", reservationResource.Resource.Name, reservationResource.Quantity ), listSubHeaderFont ) );
+                                    headerCell.Border = 0;
+                                    headerCell.Colspan = 8;
+                                    listResourceTable.AddCell( headerCell );
+
+                                    var columnCount = 0;
+
+                                    foreach ( var attributeDict in reservationResource.Attributes )
+                                    {
+                                        var resourceTableCell = new PdfPCell();
+                                        resourceTableCell.Border = 0;
+                                        resourceTableCell.Colspan = 2;
+
+                                        var attribute = attributeDict.Value;
+                                        var value = reservationResource.GetAttributeValue( attribute.Key );
+                                        var formattedValue = attribute.FieldType.Field.FormatValue( null, attribute.EntityTypeId, reservationResource.Id, value, attribute.QualifierValues, true );
+
+                                        var resourceListItem = new Phrase( attribute.Name, listItemFontBold );
+                                        var resourceListItemValue = new Phrase( formattedValue, listItemFontNormal );
+
+                                        if ( !string.IsNullOrWhiteSpace( value ) )
+                                        {
+                                            if ( columnCount == 8 )
+                                            {
+                                                columnCount = 0;
+                                            }
+                                            columnCount += 2;
+                                            addResourceTable = true;
+                                            resourceTableCell.AddElement( resourceListItem );
+                                            resourceTableCell.AddElement( resourceListItemValue );
+                                            listResourceTable.AddCell( resourceTableCell );
+                                        }
+                                    }
+                                    if ( columnCount > 0 && columnCount < 8 )
+                                    {
+                                        PdfPCell columnFillCell = new PdfPCell( new Phrase( Chunk.NEWLINE ) );
+                                        columnFillCell.Border = PdfPCell.NO_BORDER;
+                                        columnFillCell.Colspan = 8 - columnCount;
+                                        columnFillCell.FixedHeight = 15;
+
+                                        listResourceTable.AddCell( columnFillCell );
+                                    }
+
+                                    //For blank line
+                                    PdfPCell blankCell = new PdfPCell( new Phrase( Chunk.NEWLINE ) );
+                                    blankCell.Border = PdfPCell.NO_BORDER;
+                                    blankCell.Colspan = 8;
+                                    blankCell.FixedHeight = 15;
+
+                                    listResourceTable.AddCell( blankCell );
+                                }
+                            }
+
+                            if ( addResourceTable )
+                            {
+                                PdfPCell borderCell = new PdfPCell( new Phrase( Chunk.NEWLINE ) );
+                                borderCell.Colspan = 8;
+                                borderCell.FixedHeight = 5;
+                                borderCell.BorderWidth = 0;
+                                borderCell.BorderWidthBottom = 1;
+                                borderCell.BorderColorBottom = Color.DARK_GRAY;
+
+                                listResourceTable.AddCell( borderCell );
+
+                                //document.Add( listResourceTable );
+
+                                PdfPCell resourceRow = new PdfPCell();
+                                resourceRow.Border = 0;
+                                resourceRow.AddElement( listResourceTable );
+                                keepTogetherEvent.AddCell( resourceRow );
+                            }
+                        }
+                        document.Add( keepTogetherEvent );
                     }
                 }
                 document.NewPage();
@@ -301,27 +406,32 @@ namespace rocks.kfs.RoomManagement.ReportTemplates
     }
 }
 
-public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
+public class ResourceQuestionHeaderFooter : PdfPageEventHelper
 {
     // This is the contentbyte object of the writer
-    PdfContentByte cb;
+    private PdfContentByte cb;
+
     // we will put the final number of pages in a template
-    PdfTemplate template;
+    private PdfTemplate template;
+
     // this is the BaseFont we are going to use for the header / footer
-    BaseFont bf = null;
+    private BaseFont bf = null;
+
     // This keeps track of the creation time
-    DateTime PrintTime = DateTime.Now;
+    private DateTime PrintTime = DateTime.Now;
 
     #region Properties
+
     private string _Title;
+
     public string Title
     {
         get { return _Title; }
         set { _Title = value; }
     }
 
-
     private string _CalendarDate;
+
     public string CalendarDate
     {
         get { return _CalendarDate; }
@@ -329,24 +439,31 @@ public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
     }
 
     private string _HeaderLeft;
+
     public string HeaderLeft
     {
         get { return _HeaderLeft; }
         set { _HeaderLeft = value; }
     }
+
     private string _HeaderRight;
+
     public string HeaderRight
     {
         get { return _HeaderRight; }
         set { _HeaderRight = value; }
     }
+
     private Font _HeaderFont;
+
     public Font HeaderFont
     {
         get { return _HeaderFont; }
         set { _HeaderFont = value; }
     }
+
     private Font _SubHeaderFont;
+
     public Font SubHeaderFont
     {
         get { return _SubHeaderFont; }
@@ -354,13 +471,15 @@ public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
     }
 
     private bool _IsHeaderShown;
+
     public bool IsHeaderShown
     {
         get { return _IsHeaderShown; }
         set { _IsHeaderShown = value; }
     }
 
-    #endregion
+    #endregion Properties
+
     // we override the onOpenDocument method
     public override void OnOpenDocument( PdfWriter writer, Document document )
     {
@@ -386,9 +505,9 @@ public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
             document.Add( new Paragraph( CalendarDate, HeaderFont ) );
 
             //Build Subheaders
-            var listSubHeaderTable = new PdfPTable( 4 );
+            var listSubHeaderTable = new PdfPTable( 8 );
             listSubHeaderTable.LockedWidth = true;
-            listSubHeaderTable.TotalWidth = PageSize.A4.Rotate().Width - document.LeftMargin - document.RightMargin;
+            listSubHeaderTable.TotalWidth = PageSize.LETTER.Rotate().Width - document.LeftMargin - document.RightMargin;
             listSubHeaderTable.HorizontalAlignment = 0;
             listSubHeaderTable.SpacingBefore = 10;
             listSubHeaderTable.SpacingAfter = 0;
@@ -399,7 +518,7 @@ public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
             listSubHeaderTable.AddCell( new Phrase( "Name", SubHeaderFont ) );
             listSubHeaderTable.AddCell( new Phrase( "Event Time", SubHeaderFont ) );
             listSubHeaderTable.AddCell( new Phrase( "Reservation Time", SubHeaderFont ) );
-            listSubHeaderTable.AddCell( new Phrase( "Locations: NC, Unlock", SubHeaderFont ) );
+            listSubHeaderTable.AddCell( new Phrase( "Locations", SubHeaderFont ) );
             listSubHeaderTable.AddCell( new Phrase( "Resources", SubHeaderFont ) );
             listSubHeaderTable.AddCell( new Phrase( "Has Layout", SubHeaderFont ) );
             listSubHeaderTable.AddCell( new Phrase( "Status", SubHeaderFont ) );
@@ -432,6 +551,7 @@ public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
         pageSize.GetBottom( 30 ), 0 );
         cb.EndText();
     }
+
     public override void OnCloseDocument( PdfWriter writer, Document document )
     {
         base.OnCloseDocument( writer, document );
@@ -441,5 +561,4 @@ public class SPACTwoColumnHeaderFooter : PdfPageEventHelper
         template.ShowText( "" + ( writer.PageNumber - 1 ) );
         template.EndText();
     }
-
 }
